распределение методов
