+++ composer.json
@@ -20,7 +20,7 @@
             "type":"package",
             "package":{
                 "name":"domatskiy/skyparkcdn",
-                "version":"0.3.0",
+                "version":"0.5.0",
                 "source":{
                     "type":"git",
                     "url":"https://github.com/domatskiy/skyparkcdn",
