+++ src/SkyparkCDN/Method.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Domatskiy\SkyparkCDN;
+
+abstract class Method extends Request
+{
+    function __construct($token)
+    {
+        $this->token = $token;
+    }
+}
